ride_sharing/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.1.7
django-debug-toolbar==5.1.0
django-extensions==3.2.3
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
flake8==7.1.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
phonenumbers==9.0.1
pillow==11.1.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
